Refactor to make full use of material-ui styling and components
