--- src/components/Layout/Header.scss
@@ -1,19 +1,0 @@
-.header {
-  margin-bottom: 1.45rem;
-  background: rebeccapurple;
-}
-
-.header__container {
-  max-width: 960px;
-  margin: 0 auto;
-  padding: 1.45rem 1.0875rem;
-}
-
-.header__title {
-  margin: 0;
-}
-
-.header__link {
-  color: white;
-  text-decoration: none;
-}
