--- src/components/Layout/Layout.scss
@@ -1,7 +1,0 @@
-@import 'normalize';
-
-.main-layout {
-  // max-width: 960px;
-  margin: 0 auto;
-  padding: 0 1.0875rem 1.45rem;
-}
