--- src/components/Layout/normalize.scss
@@ -1,443 +1,0 @@
-// sass-lint:disable no-vendor-prefixes
-
-html {
-  box-sizing: border-box;
-  overflow-y: scroll;
-  font: 112.5%/1.45em georgia, serif;
-  -ms-text-size-adjust: 100%;
-  -webkit-text-size-adjust: 100%;
-
-  @media only screen and (max-width: 480px) {
-    font-size: 100%;
-  }
-}
-
-body {
-  margin: 0;
-  color: hsla(0, 0%, 0%, .8);
-  font-family: georgia, serif;
-  font-weight: normal;
-  word-wrap: break-word;
-  font-kerning: normal;
-  font-feature-settings: 'kern', 'liga', 'clig', 'calt';
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-article,
-aside,
-details,
-figcaption,
-figure,
-footer,
-header,
-main,
-menu,
-nav,
-section,
-summary {
-  display: block;
-}
-
-audio,
-canvas,
-progress,
-video {
-  display: inline-block;
-}
-
-audio:not([controls]) {
-  display: none;
-  height: 0;
-}
-
-progress {
-  vertical-align: baseline;
-}
-
-[hidden],
-template {
-  display: none;
-}
-
-a {
-  background-color: transparent;
-  -webkit-text-decoration-skip: objects;
-}
-
-a:active,
-a:hover {
-  outline-width: 0;
-}
-
-abbr[title] {
-  border-bottom: 1px dotted hsla(0, 0%, 0%, .5);
-  text-decoration: none;
-  cursor: help;
-}
-
-b,
-strong {
-  font-weight: inherit;
-  font-weight: bolder; // sass-lint:disable-line no-duplicate-properties
-}
-
-dfn {
-  font-style: italic;
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  margin: 0 0 1.45rem;
-  padding: 0;
-  color: inherit;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen, Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-  font-weight: bold;
-  line-height: 1.1;
-  text-rendering: optimizeLegibility;
-}
-
-h1 {
-  font-size: 2.25rem;
-}
-
-h2 {
-  font-size: 1.62671rem;
-}
-
-h3 {
-  font-size: 1.38316rem;
-}
-
-h4 {
-  font-size: 1rem;
-}
-
-h5 {
-  font-size: .85028rem;
-}
-
-h6 {
-  font-size: .78405rem;
-}
-
-mark {
-  background-color: #ff0;
-  color: #000;
-}
-
-small {
-  font-size: 80%;
-}
-
-sub,
-sup {
-  position: relative;
-  font-size: 75%;
-  line-height: 0;
-  vertical-align: baseline;
-}
-
-sub {
-  bottom: -.25em;
-}
-
-sup {
-  top: -.5em;
-}
-
-img {
-  max-width: 100%;
-  margin: 0 0 1.45rem;
-  padding: 0;
-  border-style: none;
-}
-
-svg:not(:root) {
-  overflow: hidden;
-}
-
-code,
-kbd,
-pre,
-samp {
-  font-family: monospace, monospace;
-  font-size: 1em;
-}
-
-hr {
-  box-sizing: content-box;
-  height: 1px;
-  margin: 0 0 calc(1.45rem - 1px);
-  padding: 0;
-  overflow: visible;
-  border: 0;
-  background: hsla(0, 0%, 0%, .2);
-}
-
-button,
-input,
-optgroup,
-select,
-textarea {
-  margin: 0;
-  font: inherit;
-}
-
-optgroup {
-  font-weight: 700;
-}
-
-button,
-input {
-  overflow: visible;
-}
-
-button,
-select {
-  text-transform: none;
-}
-
-[type='reset'],
-[type='submit'],
-button,
-html [type='button'] {
-  -webkit-appearance: button;
-}
-
-[type='button']::-moz-focus-inner,
-[type='reset']::-moz-focus-inner,
-[type='submit']::-moz-focus-inner,
-button::-moz-focus-inner {
-  padding: 0;
-  border-style: none;
-}
-
-[type='button']:-moz-focusring,
-[type='reset']:-moz-focusring,
-[type='submit']:-moz-focusring,
-button:-moz-focusring {
-  outline: 1px dotted ButtonText;
-}
-
-fieldset {
-  margin: 0 2px;
-  padding: .35em .625em .75em;
-  border: 1px solid silver;
-}
-
-legend {
-  display: table;
-  box-sizing: border-box;
-  max-width: 100%;
-  padding: 0;
-  color: inherit;
-  white-space: normal;
-}
-
-textarea {
-  overflow: auto;
-}
-
-[type='checkbox'],
-[type='radio'] {
-  box-sizing: border-box;
-  padding: 0;
-}
-
-[type='number']::-webkit-inner-spin-button,
-[type='number']::-webkit-outer-spin-button {
-  height: auto;
-}
-
-[type='search'] {
-  outline-offset: -2px;
-  -webkit-appearance: textfield;
-}
-
-[type='search']::-webkit-search-cancel-button,
-[type='search']::-webkit-search-decoration {
-  -webkit-appearance: none;
-}
-
-::-webkit-input-placeholder {
-  opacity: .54;
-  color: inherit;
-}
-
-::-webkit-file-upload-button {
-  font: inherit;
-  -webkit-appearance: button;
-}
-
-* {
-  box-sizing: inherit;
-}
-
-*:before {
-  box-sizing: inherit;
-}
-
-*:after {
-  box-sizing: inherit;
-}
-
-hgroup,
-dl,
-dd,
-p,
-figure,
-table,
-fieldset,
-form,
-noscript,
-iframe,
-address {
-  margin: 0 0 1.45rem;
-  padding: 0;
-}
-
-ul,
-ol {
-  margin: 0 0 1.45rem 1.45rem;
-  padding: 0;
-  list-style-position: outside;
-  list-style-image: none;
-}
-
-pre {
-  margin: 0 0 1.45rem;
-  padding: 1.45rem;
-  overflow: auto;
-  border-radius: 3px;
-  background: hsla(0, 0%, 0%, .04);
-  font-size: .85rem;
-  line-height: 1.42;
-  word-wrap: normal;
-}
-
-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 1rem;
-  line-height: 1.45rem;
-}
-
-blockquote {
-  margin: 0 1.45rem 1.45rem;
-  padding: 0;
-}
-
-b,
-strong,
-dt,
-th {
-  font-weight: bold;
-}
-
-li {
-  margin-bottom: calc(1.45rem / 2);
-
-  ol &,
-  ul & {
-    padding-left: 0;
-  }
-
-  > ol,
-  > ul {
-    margin-top: calc(1.45rem / 2);
-    margin-bottom: calc(1.45rem / 2);
-    margin-left: 1.45rem;
-  }
-
-  > p {
-    margin-bottom: calc(1.45rem / 2);
-  }
-}
-
-blockquote,
-li,
-p {
-  *:last-child {
-    margin-bottom: 0;
-  }
-}
-
-code {
-  font-size: .85rem;
-  line-height: 1.45rem;
-}
-
-kbd {
-  font-size: .85rem;
-  line-height: 1.45rem;
-}
-
-samp {
-  font-size: .85rem;
-  line-height: 1.45rem;
-}
-
-abbr {
-  border-bottom: 1px dotted hsla(0, 0%, 0%, .5);
-  cursor: help;
-}
-
-acronym {
-  border-bottom: 1px dotted hsla(0, 0%, 0%, .5);
-  cursor: help;
-}
-
-thead {
-  text-align: left;
-}
-
-td,
-th {
-  padding: .725rem .966667rem calc(.725rem - 1px);
-  border-bottom: 1px solid hsla(0, 0%, 0%, .12);
-  text-align: left;
-  font-feature-settings: 'tnum';
-}
-
-th:first-child,
-td:first-child {
-  padding-left: 0;
-}
-
-th:last-child,
-td:last-child {
-  padding-right: 0;
-}
-
-tt,
-code {
-  padding: .2em;
-  border-radius: 3px;
-  background-color: hsla(0, 0%, 0%, .04);
-  font-family: 'SFMono-Regular', Consolas, 'Roboto Mono', 'Droid Sans Mono', 'Liberation Mono', Menlo, Courier, monospace;
-}
-
-pre code {
-  background: none;
-  line-height: 1.42;
-}
-
-code:before,
-code:after,
-tt:before,
-tt:after {
-  content: ' ';
-  letter-spacing: -.2em;
-}
-
-pre code:before,
-pre code:after,
-pre tt:before,
-pre tt:after {
-  content: '';
-}
