+++ src/components/Shops/Card.jsx
@@ -5,40 +5,31 @@ import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
 import CardActions from '@material-ui/core/CardActions';
 import CardContent from '@material-ui/core/CardContent';
-import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 
-const useStyles = makeStyles({
-  root: {
-    flexGrow: 1,
-  },
-});
+import withClasses from 'Styles/components/shops/card';
 
-const ShopsCard = ({ shop }) => {
-  const classes = useStyles();
+const ShopsCard = ({ classes, shop }) => (
+  <Card className={ classes.root }>
+    <CardActionArea>
+      <CardContent>
+        <Typography gutterBottom variant="h5" component="h2">
+          { shop.nome }
+        </Typography>
+        <Typography variant="body2" color="textSecondary" component="p">
+          { shop.oferta }
+        </Typography>
+      </CardContent>
+    </CardActionArea>
+    <CardActions>
+      <Button size="small" color="primary">
+        { shop.localiza__o__distrito_ }
+      </Button>
+      <Button size="small" color="primary">
+        Learn More
+      </Button>
+    </CardActions>
+  </Card>
+);
 
-  return (
-    <Card className={ classes.root }>
-      <CardActionArea>
-        <CardContent>
-          <Typography gutterBottom variant="h5" component="h2">
-            { shop.nome }
-          </Typography>
-          <Typography variant="body2" color="textSecondary" component="p">
-            { shop.oferta }
-          </Typography>
-        </CardContent>
-      </CardActionArea>
-      <CardActions>
-        <Button size="small" color="primary">
-          { shop.localiza__o__distrito_ }
-        </Button>
-        <Button size="small" color="primary">
-          Learn More
-        </Button>
-      </CardActions>
-    </Card>
-  );
-};
-
-export default ShopsCard;
+export default withClasses(ShopsCard);
