+++ src/pages/404.jsx
@@ -1,14 +1,14 @@
 import React from 'react';
 
-import Layout from '../components/Layout/Layout';
+import PageWrapper from 'Layouts/PageWrapper';
 import SEO from '../components/SEO';
 
 const NotFoundPage = () => (
-  <Layout>
+  <PageWrapper>
     <SEO title="404: Not found" />
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </Layout>
+  </PageWrapper>
 );
 
 export default NotFoundPage;
