+++ src/polyfills/string-normalize.js
