+++ src/styles/components/shops/card.js
@@ -1,0 +1,9 @@
+import wrapStyles from 'Styles/wrapStyles';
+
+const styles = {
+  root: {
+    flexGrow: 1,
+  },
+};
+
+export default component => wrapStyles(styles, component);
