+++ src/styles/layouts/pageWrapper.js
@@ -1,0 +1,10 @@
+import wrapStyles from 'Styles/wrapStyles';
+
+const styles = {
+  root: {
+    margin: '0 auto',
+    padding: 0,
+  },
+};
+
+export default component => wrapStyles(styles, component);
