+++ src/styles/wrapStyles.js
@@ -1,0 +1,3 @@
+import { withStyles } from '@material-ui/core/styles';
+
+export default (styles, component) => withStyles(styles)(component);
